refactor(themes): share organization design package validation
